+++ src/main/java/cn/yuqi/mini/spring/beans/factory/BeanDefinition.java
@@ -1,0 +1,11 @@
+package cn.yuqi.mini.spring.beans.factory;
+
+public class BeanDefinition {
+    private Class<?> beanClass;
+    public BeanDefinition(Class<?> beanClass) {
+        this.beanClass = beanClass;
+    }
+    public Class<?> getBeanClass(){
+        return this.beanClass;
+    }
+}