+++ src/main/java/cn/yuqi/mini/spring/beans/factory/BeanFactory.java
@@ -1,5 +1,5 @@
 package cn.yuqi.mini.spring.beans.factory;
 
 public interface BeanFactory {
-    Object getBean(String name);
+    Object getBean(String beanName);
 }